+++ packages/app/studio/src/opendaw-api/OpenPresetAPI.ts
@@ -2,10 +2,9 @@ import {
     asDefined, DefaultObservableValue, isDefined, Lazy, panic, Procedure, RuntimeNotifier, tryCatch, unitValue, UUID
 } from "@opendaw/lib-std"
 import {network, Promises} from "@opendaw/lib-runtime"
-import {base64Credentials, OpenDAWHeaders} from "../OpenDAWHeaders"
-import {PresetMeta} from "./PresetMeta"
+import {base64Credentials, OpenDAWHeaders} from "./OpenDAWHeaders"
+import {PresetMeta} from "@opendaw/studio-core"
 
-/** @internal */
 export class OpenPresetAPI {
     static readonly ApiRoot = "https://api.opendaw.studio/presets"
     static readonly FileRoot = "https://assets.opendaw.studio/presets"