+++ packages/studio/core/src/cloud/CloudBackupSamples.ts
@@ -2,7 +2,8 @@ import {Arrays, Errors, panic, Procedure, Progress, RuntimeNotifier, UUID} from 
 import {network, Promises} from "@opendaw/lib-runtime"
 import {SamplePeaks} from "@opendaw/lib-fusion"
 import {Sample} from "@opendaw/studio-adapters"
-import {OpenSampleAPI, SampleStorage} from "../samples"
+import {SampleStorage} from "../samples"
+import {FactoryCatalog} from "../FactoryCatalog"
 import {CloudHandler} from "./CloudHandler"
 import {Workers} from "../Workers"
 import {WavFile} from "@opendaw/lib-dsp"
@@ -21,7 +22,7 @@ export class CloudBackupSamples {
                        log: Procedure<string>) {
         log("Collecting all sample domains...")
         const [stock, local, cloud] = await Promise.all([
-            OpenSampleAPI.get().all(),
+            FactoryCatalog.get().samples(),
             SampleStorage.get().list(),
             cloudHandler.download(CloudBackupSamples.RemoteCatalogPath)
                 .then(json => JSON.parse(new TextDecoder().decode(json)))