+++ packages/studio/core/src/samples/SampleService.ts
@@ -10,7 +10,7 @@ import {FilePickerAcceptTypes} from "../FilePickerAcceptTypes"
 import {WavFile} from "@opendaw/lib-dsp"
 import {Workers} from "../Workers"
 import {SampleStorage} from "./SampleStorage"
-import {OpenSampleAPI} from "./OpenSampleAPI"
+import {FactoryCatalog} from "../FactoryCatalog"
 
 export class SampleService extends AssetService<Sample, AudioData> {
     protected readonly namePlural: string = "Samples"
@@ -61,7 +61,7 @@ export class SampleService extends AssetService<Sample, AudioData> {
     }
 
     protected async collectAllFiles(): Promise<ReadonlyArray<Sample>> {
-        const stock = await OpenSampleAPI.get().all()
+        const stock = await FactoryCatalog.get().samples()
         const local = await SampleStorage.get().list()
         return Arrays.merge(stock, local, (sample, {uuid}) => sample.uuid === uuid)
     }