+++ packages/lib/std/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.79](https://github.com/andremichelle/openDAW/compare/@opendaw/lib-std@0.0.78...@opendaw/lib-std@0.0.79) (2026-06-23)
+
+**Note:** Version bump only for package @opendaw/lib-std
+
 ## [0.0.78](https://github.com/andremichelle/openDAW/compare/@opendaw/lib-std@0.0.77...@opendaw/lib-std@0.0.78) (2026-05-22)
 
 **Note:** Version bump only for package @opendaw/lib-std